+++ Examples/APIBlox Features/Examples.Features/Controllers/ExamplesController.cs
@@ -119,14 +119,15 @@ namespace Examples.Controllers
         [HttpGet("serverFault")]
         public ActionResult ThrowExceptionForServerFaultExample(string exceptionMessage)
         {
-            throw new Exception("Be sure to try this out in RELEASE mode",
-                new IndexOutOfRangeException("As most if not all of this",
-                    new ArgumentException("error information is NOT displayed in production",
-                        new FileNotFoundException("By the way here is your message",
-                            new NullReferenceException(exceptionMessage)
-                        )
-                    )
-                )
+            throw new Exception("Be sure to try this out in RELEASE mode"
+                //,
+                //new IndexOutOfRangeException("As most if not all of this",
+                //    new ArgumentException("error information is NOT displayed in production",
+                //        new FileNotFoundException("By the way here is your message",
+                //            new NullReferenceException(exceptionMessage)
+                //        )
+                //    )
+                //)
             );
         }